ml/requirements.txt
pandas
numpy
elasticsearch==8.6.2
scikit-surprise
matplotlib
scikit-learn
jupyter
faiss-cpu
fuzzywuzzy
sentence_transformers